TS: EntryMenuLink
